tools/travelogue-generator/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-genai
python-dotenv
pillow
requests
